+++ src/product-share/containers/ValidateShareForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Relay from 'react-relay';
 import ActionPanelForm from 'shared/components/ActionPanelForm';
-import {TextInput} from 'shared/components/Form';
+import {TextInput, validationErrors} from 'shared/components/Form';
 
 class Container extends React.Component {
   static propTypes = {
@@ -63,6 +63,7 @@ class Container extends React.Component {
           label={'Token'}
           validations={'isExisty'}
           value={isFarmOwner ? share.token : ''}
+          validationErrors={validationErrors.token}
           required
         />
       </ActionPanelForm>;